+++ js/config.js
@@ -1,14 +1,14 @@
 /*
 ==================================================
 RosalitaRP Explorer
-Version : 1.3.29
+Version : 1.3.30
 Creator : Rathan
 ==================================================
 */
 
 const APP_NAME = "RosalitaRP Explorer";
 const APP_SUBTITLE = "Community Resource Map";
-const APP_VERSION = "1.3.29";
+const APP_VERSION = "1.3.30";
 const APP_CREATOR = "Rathan";
 const DEFAULT_DANGER_RADIUS = 50;